requirements.txt
Flask
Flask-BasicAuth
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
marshmallow
marshmallow-sqlalchemy
python-dotenv
